requirements.txt
Flask
requests
selenium
python-dotenv
pymongo
